Use STATE_CLASS_TOTAL with last_reset for everything
Neither STATE_CLASS_TOTAL_INCREASING nor STATE_CLASS_MEASUREMENT work
well for the ca. 9am reset that seems to happen for daily summary data.

The `last_reset` time is set to whenever a value changes. This means
this is liable to a bug where a value will be missed it it is the same
as the previous day.

Signed-off-by: Olivier Mehani <shtrom@ssji.net>